Day7

+++ src/main/kotlin/Day1.kt
@@ -3,16 +3,16 @@ import kotlin.io.path.readLines
 import kotlin.math.abs
 
 class Day1: Day(1) {
-    override fun part1(input: Path): Int {
+    override fun part1(input: Path): Long {
         val (list1, list2) = input.readLines().toIntPairs().unzip()
         return list1.sorted()
             .zip(list2.sorted())
-            .sumOf { (x, y) -> abs(x - y) }
+            .sumOf { (x, y) -> abs(x - y) }.toLong()
     }
 
-    override fun part2(input: Path): Int {
+    override fun part2(input: Path): Long {
         val (list1, list2) = input.readLines().toIntPairs().unzip()
         val count = list1.groupingBy { it }.eachCount()
-        return list2.sumOf { it * count.getOrDefault(it, 0) }
+        return list2.sumOf { it * count.getOrDefault(it, 0) }.toLong()
     }
 }
